requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.124.2
flatbuffers==25.9.23
h11==0.16.0
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
piper-tts==1.3.0
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
requests==2.32.5
sounddevice==0.5.3
soundfile==0.13.1
starlette==0.50.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
